update on delete functions

+++ app/Http/Controllers/CitiesController.php
@@ -5,12 +5,15 @@ namespace App\Http\Controllers;
 use App\Models\City;
 
 use App\Http\Requests\CityRequest;
+use App\Models\BuyPackage;
 use App\Models\Gym;
 use App\Models\Session;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Yajra\DataTables\Facades\DataTables;
 use Illuminate\Database\QueryException;
+use Illuminate\Support\Facades\Redirect;
 
 class CitiesController extends Controller
 {
@@ -52,12 +55,17 @@ class CitiesController extends Controller
 
     public function destroy($cityID)
     {
-        try {
+        $checkGym = Gym::where('city_id', $cityID)->first();
+        $checkUserORManager = User::where('city_id', $cityID)->first();
+        $checkBuyPackage = BuyPackage::where('city_id', $cityID)->first();
+
+        if ($checkUserORManager == null && $checkBuyPackage == null && $checkGym == null) {
             City::findOrFail($cityID)->delete();
-            return redirect()->route('cities.index');
-        } catch (QueryException $e) {
-            return redirect()->route('cities.index');
+            return to_route('cities.index')->with('success', 'City deleted successfully');
+        } else {
+            return Redirect::back()->withErrors(['message' => 'delete']);
         }
+
     }
 
     public function deleteMedia($oldImg, $path)
